# backend/domains/ops/evidence/src/controllers/evidence-access-request.controller.ts
// Generated from BP-OPS-EVIDENCE-001 v1.0.0 sha256:a8692e7ee4171aea45d3aa6a8ca457251f3005b1dc05d1b03e2aa5ebc8f1923f
import {
  Body,
  Controller,
  Delete,
  Get,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import { Action, Audit, Resource } from '@detran/shared';
import type { CreateEvidenceAccessRequestDto } from '../dto/create-evidence-access-request.dto.js';
import { EvidenceAccessRequestService } from '../services/evidence-access-request.service.js';

@Controller('v1/ops/evidence/evidence-access-requests')
@Resource('ops:evidence-access-request')
export class EvidenceAccessRequestController {
  constructor(private readonly service: EvidenceAccessRequestService) {}
  @Get() @Action('read') list() {
    return this.service.findAll();
  }
  @Get(':id') @Action('read') get(@Param('id') id: string) {
    return this.service.findOne(id);
  }
  @Post()
  @Action('create')
  @Audit({
    action: 'OPS_EVIDENCE_ACCESS_REQUEST_CREATE',
    entity: 'ops.evidence_access_request',
  })
  create(@Body() dto: CreateEvidenceAccessRequestDto) {
    return this.service.create(dto);
  }
  @Patch(':id')
  @Action('update')
  @Audit({
    action: 'OPS_EVIDENCE_ACCESS_REQUEST_UPDATE',
    entity: 'ops.evidence_access_request',
  })
  update(
    @Param('id') id: string,
    @Body() dto: Partial<CreateEvidenceAccessRequestDto>,
  ) {
    return this.service.update(id, dto);
  }
  @Delete(':id')
  @Action('delete')
  @Audit({
    action: 'OPS_EVIDENCE_ACCESS_REQUEST_DELETE',
    entity: 'ops.evidence_access_request',
  })
  remove(@Param('id') id: string) {
    return this.service.remove(id);
  }
}